+++ tests/unit/private/Get-DscResourceHelpExampleContent.Tests.ps1
@@ -328,7 +328,7 @@ Configuration Example
 .COMPANYNAME Microsoft Corporation
 .COPYRIGHT
 .TAGS DSCConfiguration
-.LICENSEURI https://github.com/PowerShell/CertificateDsc/blob/master/LICENSE
+.LICENSEURI https://github.com/PowerShell/CertificateDsc/blob/main/LICENSE
 .PROJECTURI https://github.com/PowerShell/CertificateDsc
 .ICONURI
 .EXTERNALMODULEDEPENDENCIES
@@ -414,7 +414,7 @@ Configuration Example
 .COMPANYNAME Microsoft Corporation
 .COPYRIGHT
 .TAGS DSCConfiguration
-.LICENSEURI https://github.com/PowerShell/CertificateDsc/blob/master/LICENSE
+.LICENSEURI https://github.com/PowerShell/CertificateDsc/blob/main/LICENSE
 .PROJECTURI https://github.com/PowerShell/CertificateDsc
 .ICONURI
 .EXTERNALMODULEDEPENDENCIES
@@ -577,7 +577,7 @@ Configuration CertificateExport_CertByFriendlyName_Config
 .COMPANYNAME Microsoft Corporation
 .COPYRIGHT
 .TAGS DSCConfiguration
-.LICENSEURI https://github.com/PowerShell/CertificateDsc/blob/master/LICENSE
+.LICENSEURI https://github.com/PowerShell/CertificateDsc/blob/main/LICENSE
 .PROJECTURI https://github.com/PowerShell/CertificateDsc
 .ICONURI
 .EXTERNALMODULEDEPENDENCIES