+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Renamed default branch to `main` ([issue #49](https://github.com/dsccommunity/DscResource.DocGenerator/issues/49)).
+
 ## [0.7.2] - 2021-01-17
 
 ### Fixed